requirements.txt
fastapi
uvicorn[standard]
websockets
python-dotenv
audioop-lts; python_version >= '3.13'
